Move ModelAliasDefinition class into model package
